+++ cloud/testserver/deployments.go
@@ -91,8 +91,11 @@ func PostDeployment(store *cloudstore.Data, w http.ResponseWriter, r *http.Reque
 	var stackIDs []int64
 	res := cloud.DeploymentStacksResponse{}
 	for _, s := range rPayload.Stacks {
+		if s.Stack.Target == "" {
+			s.Stack.Target = "default"
+		}
 		state := cloudstore.NewState()
-		gotStack, _, found := store.GetStackByMetaID(org, s.Stack.MetaID)
+		gotStack, _, found := store.GetStackByMetaID(org, s.Stack.MetaID, s.Stack.Target)
 		if found {
 			state = gotStack.State
 		}