+++ e2etests/core/generate_test.go
@@ -25,7 +25,7 @@ import (
 // changing so testing this in an e2e manner makes it less liable to
 // break because of structural changes.
 
-func TestGenerate(t *testing.T) {
+func TestE2EGenerate(t *testing.T) {
 	t.Parallel()
 
 	type (
@@ -42,6 +42,7 @@ func TestGenerate(t *testing.T) {
 		testcase struct {
 			name             string
 			layout           []string
+			fromdir          string
 			files            []file
 			detailedExitCode bool
 			want             want
@@ -80,6 +81,19 @@ func TestGenerate(t *testing.T) {
 				"s:stacks/stack-1",
 				"s:stacks/stack-2",
 			},
+			want: want{
+				run: RunExpected{
+					Stdout: noCodegenMsg,
+				},
+			},
+		},
+		{
+			name: "stacks with no codegen - run from dir",
+			layout: []string{
+				"s:stacks/stack-1",
+				"s:stacks/stack-2",
+			},
+			fromdir: "/stacks",
 			want: want{
 				run: RunExpected{
 					Stdout: noCodegenMsg,
@@ -145,6 +159,36 @@ Hint: '+', '~' and '-' mean the file was created, changed and deleted, respectiv
 						path: p("/stack/file.txt"),
 						body: str("hi"),
 					},
+				},
+			},
+		},
+		{
+			name: "generate file and hcl - from empty dir",
+			layout: []string{
+				"s:stack",
+				"d:empty",
+			},
+			fromdir: "/empty",
+			files: []file{
+				{
+					path: p("/config.tm"),
+					body: Doc(
+						GenerateHCL(
+							Labels("file.hcl"),
+							Content(
+								Str("a", "hi"),
+							),
+						),
+						GenerateFile(
+							Labels("file.txt"),
+							Str("content", "hi"),
+						),
+					),
+				},
+			},
+			want: want{
+				run: RunExpected{
+					Stdout: noCodegenMsg,
 				},
 			},
 		},
@@ -418,7 +462,7 @@ Hint: '+', '~' and '-' mean the file was created, changed and deleted, respectiv
 				)
 			}
 
-			tmcli := NewCLI(t, s.RootDir())
+			tmcli := NewCLI(t, filepath.Join(s.RootDir(), tcase.fromdir))
 			args := []string{"generate"}
 			if tcase.detailedExitCode {
 				args = append(args, "--detailed-exit-code")
@@ -447,30 +491,8 @@ Hint: '+', '~' and '-' mean the file was created, changed and deleted, respectiv
 	}
 }
 
-func TestGenerateIgnoresWorkingDirectory(t *testing.T) {
+func TestE2EGenerateRespectsWorkingDirectory(t *testing.T) {
 	t.Parallel()
-	wantStdout := generate.Report{
-		Successes: []generate.Result{
-			{
-				Dir: project.NewPath("/"),
-				Created: []string{
-					"root.stacks.txt",
-				},
-			},
-			{
-				Dir: project.NewPath("/stacks/stack-1"),
-				Created: []string{
-					"stack.hcl", "stack.name.txt",
-				},
-			},
-			{
-				Dir: project.NewPath("/stacks/stack-2"),
-				Created: []string{
-					"stack.hcl", "stack.name.txt",
-				},
-			},
-		},
-	}.Full() + "\n"
 
 	configStr := Doc(
 		GenerateFile(
@@ -491,7 +513,7 @@ func TestGenerateIgnoresWorkingDirectory(t *testing.T) {
 		),
 	).String()
 
-	runFromDir := func(t *testing.T, wd string) {
+	runFromDir := func(t *testing.T, wd string, want generate.Report) {
 		t.Run(fmt.Sprintf("terramate -C %s generate", wd), func(t *testing.T) {
 			t.Parallel()
 			s := sandbox.NoGit(t, true)
@@ -508,15 +530,60 @@ func TestGenerateIgnoresWorkingDirectory(t *testing.T) {
 			tmcli := NewCLI(t, filepath.Join(s.RootDir(), wd))
 			res := tmcli.Run("generate")
 			expected := RunExpected{
-				Stdout: wantStdout,
+				Stdout: nljoin(want.Full()),
 			}
 			AssertRunResult(t, res, expected)
 		})
 	}
 
-	runFromDir(t, "/")
-	runFromDir(t, "/stacks")
-	runFromDir(t, "/stacks/stack-1")
+	runFromDir(t, "/", generate.Report{
+		Successes: []generate.Result{
+			{
+				Dir: project.NewPath("/"),
+				Created: []string{
+					"root.stacks.txt",
+				},
+			},
+			{
+				Dir: project.NewPath("/stacks/stack-1"),
+				Created: []string{
+					"stack.hcl", "stack.name.txt",
+				},
+			},
+			{
+				Dir: project.NewPath("/stacks/stack-2"),
+				Created: []string{
+					"stack.hcl", "stack.name.txt",
+				},
+			},
+		},
+	})
+	runFromDir(t, "/stacks", generate.Report{
+		Successes: []generate.Result{
+			{
+				Dir: project.NewPath("/stacks/stack-1"),
+				Created: []string{
+					"stack.hcl", "stack.name.txt",
+				},
+			},
+			{
+				Dir: project.NewPath("/stacks/stack-2"),
+				Created: []string{
+					"stack.hcl", "stack.name.txt",
+				},
+			},
+		},
+	})
+	runFromDir(t, "/stacks/stack-1", generate.Report{
+		Successes: []generate.Result{
+			{
+				Dir: project.NewPath("/stacks/stack-1"),
+				Created: []string{
+					"stack.hcl", "stack.name.txt",
+				},
+			},
+		},
+	})
 }
 
 type str string